% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AltExpFeatureAssayPlot.R
\name{AltExpFeatureAssayPlot-class}
\alias{AltExpFeatureAssayPlot-class}
\title{AltExpFeatureAssayPlot: feature assay plot for alternative experiments}
\description{
An S4 class extending \code{\linkS4class{FeatureAssayPlot}} to plot
assay values from a feature in an \emph{alternative experiment}
(\code{\link[SingleCellExperiment]{altExp}}) stored inside a
\code{\linkS4class{SingleCellExperiment}}.
}
\details{
The Y-axis feature is typically received from an
\code{\linkS4class{AltExpRowDataTable}} panel via \code{YAxisFeatureSource},
so that clicking a row in the table immediately plots the corresponding
altExp feature across all samples.
}
\section{Slots}{

\describe{
\item{\code{AltExp}}{\code{character(1)}.  Name of the alternative experiment to
visualise (must be present in \code{altExpNames(se)}).
The \code{Assay} slot (inherited from \code{\linkS4class{FeatureAssayPlot}})
refers to an assay within this altExp.}
}}

\section{Inherited slots}{

All slots from \code{\linkS4class{FeatureAssayPlot}} are inherited,
including \code{Assay}, \code{XAxis}, \code{XAxisColumnData},
\code{YAxisFeatureName}, \code{YAxisFeatureSource}, and those controlling
colour, shape, size, faceting, and multi-selection behaviour.
}

\seealso{
\code{\link{AltExpFeatureAssayPlot}} for the constructor.
\code{\linkS4class{FeatureAssayPlot}} for the parent class.
\code{\linkS4class{AltExpRowDataTable}} for a compatible row-selection source.
}
