% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AltExpFeatureAssayPlot.R
\name{AltExpFeatureAssayPlot}
\alias{AltExpFeatureAssayPlot}
\title{Construct an AltExpFeatureAssayPlot panel}
\usage{
AltExpFeatureAssayPlot(...)
}
\arguments{
\item{...}{Named arguments corresponding to slots of
\code{\linkS4class{AltExpFeatureAssayPlot}} or its parent classes.}
}
\value{
An \code{\linkS4class{AltExpFeatureAssayPlot}} object.
}
\description{
Creates an instance of \code{\linkS4class{AltExpFeatureAssayPlot}} for use
as a panel in an iSEE application.
}
\seealso{
\code{\linkS4class{AltExpFeatureAssayPlot}} for slot details.
}
