% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AltExpReducedDimensionPlot.R
\name{AltExpReducedDimensionPlot}
\alias{AltExpReducedDimensionPlot}
\title{Construct an AltExpReducedDimensionPlot panel}
\usage{
AltExpReducedDimensionPlot(...)
}
\arguments{
\item{...}{Named arguments corresponding to slots of
\code{\linkS4class{AltExpReducedDimensionPlot}} or its parent classes.}
}
\value{
An \code{\linkS4class{AltExpReducedDimensionPlot}} object.
}
\description{
Creates an instance of \code{\linkS4class{AltExpReducedDimensionPlot}} for
use as a panel in an iSEE application.
}
\seealso{
\code{\linkS4class{AltExpReducedDimensionPlot}} for slot details.
}
