% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AltExpRowDataTable.R
\name{AltExpRowDataTable}
\alias{AltExpRowDataTable}
\title{Construct an AltExpRowDataTable panel}
\usage{
AltExpRowDataTable(...)
}
\arguments{
\item{...}{Named arguments corresponding to slots of
\code{\linkS4class{AltExpRowDataTable}} or its parent classes.}
}
\value{
An \code{\linkS4class{AltExpRowDataTable}} object.
}
\description{
Creates an instance of \code{\linkS4class{AltExpRowDataTable}} for use as
a panel in an iSEE application.
}
\seealso{
\code{\linkS4class{AltExpRowDataTable}} for slot details.
}
