% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AltExpVolcanoPlot.R
\name{AltExpVolcanoPlot}
\alias{AltExpVolcanoPlot}
\title{Construct an AltExpVolcanoPlot panel}
\usage{
AltExpVolcanoPlot(...)
}
\arguments{
\item{...}{Named arguments corresponding to slots of
\code{\linkS4class{AltExpVolcanoPlot}} or its parent classes.}
}
\value{
An \code{\linkS4class{AltExpVolcanoPlot}} object.
}
\description{
Creates an instance of \code{\linkS4class{AltExpVolcanoPlot}} for use as a
panel in an iSEE application.
}
\seealso{
\code{\linkS4class{AltExpVolcanoPlot}} for slot details.
}
