% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AltFeatureAssayPlot.R
\name{AltFeatureAssayPlot}
\alias{AltFeatureAssayPlot}
\title{Construct an AltFeatureAssayPlot panel}
\usage{
AltFeatureAssayPlot(...)
}
\arguments{
\item{...}{Named arguments corresponding to slots of
\code{\linkS4class{AltFeatureAssayPlot}} or its parent classes.}
}
\value{
An \code{\linkS4class{AltFeatureAssayPlot}} object.
}
\description{
Creates an instance of \code{\linkS4class{AltFeatureAssayPlot}} for use
as a panel in an iSEE application.
}
\seealso{
\code{\linkS4class{AltFeatureAssayPlot}} for slot details.
}
