% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AltReducedDimensionPlot.R
\name{AltReducedDimensionPlot}
\alias{AltReducedDimensionPlot}
\title{Construct an AltReducedDimensionPlot panel}
\usage{
AltReducedDimensionPlot(...)
}
\arguments{
\item{...}{Named arguments corresponding to slots of
\code{\linkS4class{AltReducedDimensionPlot}} or its parent classes.}
}
\value{
An \code{\linkS4class{AltReducedDimensionPlot}} object.
}
\description{
Creates an instance of \code{\linkS4class{AltReducedDimensionPlot}} for
use as a panel in an iSEE application.
}
\seealso{
\code{\linkS4class{AltReducedDimensionPlot}} for slot details.
}
