% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AltRowDataTable.R
\name{AltRowDataTable}
\alias{AltRowDataTable}
\title{Construct an AltRowDataTable panel}
\usage{
AltRowDataTable(...)
}
\arguments{
\item{...}{Named arguments corresponding to slots of
\code{\linkS4class{AltRowDataTable}} or its parent classes.}
}
\value{
An \code{\linkS4class{AltRowDataTable}} object.
}
\description{
Creates an instance of \code{\linkS4class{AltRowDataTable}} for use as
a panel in an iSEE application.
}
\seealso{
\code{\linkS4class{AltRowDataTable}} for slot details.
}
