% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AltVolcanoPlot.R
\name{AltVolcanoPlot}
\alias{AltVolcanoPlot}
\title{Construct an AltVolcanoPlot panel}
\usage{
AltVolcanoPlot(...)
}
\arguments{
\item{...}{Named arguments corresponding to slots of
\code{\linkS4class{AltVolcanoPlot}} or its parent classes.}
}
\value{
An \code{\linkS4class{AltVolcanoPlot}} object.
}
\description{
Creates an instance of \code{\linkS4class{AltVolcanoPlot}} for use as a
panel in an iSEE application.
}
\seealso{
\code{\linkS4class{AltVolcanoPlot}} for slot details.
}
