% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AltExpPlot.R
\name{.addDotPlotDataColor,AltExpPlot-method}
\alias{.addDotPlotDataColor,AltExpPlot-method}
\title{############################################################
# To control visual parameters
############################################################}
\usage{
\S4method{.addDotPlotDataColor}{AltExpPlot}(x, envir)
}
\description{
# Copied from DotPlot and modified on one line:  .add_extra_aesthetic_columns_AltExp
}
\details{
#' @export
setMethod(".generateOutput", "AltExpPlot", function(x, se, all_memory, all_contents) {
  # Initialize an environment storing information for generating ggplot commands
  plot_env <- new.env()
  plot_env$se <- se
  plot_env$colormap <- iSEE:::.get_colormap(se)

  all_cmds <- list()
  all_labels <- list()

  # Doing this first so that .generateDotPlotData can respond to the selection.
  all_cmds$select <- iSEE:::.processMultiSelections(x, all_memory, all_contents, plot_env)

  xy_out <- iSEE:::.generateDotPlotData(x, plot_env)
  all_cmds$xy <- xy_out$commands
  all_labels <- c(all_labels, xy_out$labels)

  # Line below is the only change compared to the DotPlot method to ensure that the aesthetics dimensions are correct
  extra_out <- .add_extra_aesthetic_columns_AltExp(x, plot_env)
  all_cmds <- c(all_cmds, extra_out$commands)
  all_labels <- c(all_labels, extra_out$labels)

  select_out2 <- iSEE:::.add_selectby_column(x, plot_env)
  all_cmds <- c(all_cmds, select_out2)

  # We need to set up the plot type before downsampling,
  # to ensure the X/Y jitter is correctly computed.
  all_cmds$setup <-  iSEE:::.choose_plot_type(plot_env)

  # Also collect the plot coordinates before downsampling.
  panel_data <- plot_env$plot.data

  # Non-data-related fiddling to affect the visual display.
  # First, scrambling the plot.data to avoid biases.
  scramble_cmds <- c(
    "# Avoid visual biases from default ordering by shuffling the points",
    sprintf("set.seed(%i);", nrow(panel_data)), # Using a deterministically different seed to keep things exciting.
    "plot.data <- plot.data[sample(nrow(plot.data)),,drop=FALSE];"
  )
  iSEE:::.textEval(scramble_cmds, plot_env)
  all_cmds$shuffle <- scramble_cmds

  # Next, reordering by priority (this is stable so any ordering due to the
  # shuffling above is still preserved within each priority level).
  priority_out <- iSEE:::.prioritizeDotPlotData(x, plot_env)
  rescaled_res <- FALSE
  if (has_priority <- !is.null(priority_out)) {
    order_cmds <- "plot.data <- plot.data[order(.priority),,drop=FALSE];"
    iSEE:::.textEval(order_cmds, plot_env)
    all_cmds$priority <- c(priority_out$commands, order_cmds)
    rescaled_res <- priority_out$rescaled
  }
  # Finally, the big kahuna of downsampling.
  all_cmds$downsample <- iSEE:::.downsample_points(x, plot_env, priority=has_priority, rescaled=rescaled_res)

  plot_out <- iSEE:::.generateDotPlot(x, all_labels, plot_env)
  all_cmds$plot <- plot_out$commands

  list(commands=all_cmds, contents=panel_data, plot=plot_out$plot, varname="plot.data")

})
}
