% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinkedFeaturesAssayPlot.R
\name{LinkedFeaturesAssayPlot}
\alias{LinkedFeaturesAssayPlot}
\title{Construct an LinkedFeaturesAssayPlot panel}
\usage{
LinkedFeaturesAssayPlot(...)
}
\arguments{
\item{...}{Named arguments corresponding to slots of
\code{\linkS4class{LinkedFeaturesAssayPlot}} or its parent classes.}
}
\value{
An \code{\linkS4class{LinkedFeaturesAssayPlot}} object.
}
\description{
Creates an instance of \code{\linkS4class{LinkedFeaturesAssayPlot}} for use as a
panel in an iSEE application.
}
\examples{
\dontrun{
# Minimal panel using runtime defaults
LinkedFeaturesAssayPlot()


# Select a protein from the main experiment and visualise its peptides
LinkedFeaturesAssayPlot(
  SelectionExperiment    = "(Main)",
  LookupColumn       = "Proteins",
  Experiment            = "peptides",
  AltAssay           = "peptides_norm",
  MapColumn          = "Proteins"
)

# Select a peptide from an altExp and visualise its protein in the main
# experiment
LinkedFeaturesAssayPlot(
  SelectionExperiment    = "peptides",
  LookupColumn       = "Proteins",
  Experiment            = "(Main)",
  AltAssay           = "proteins",
  MapColumn          = "Proteins"
)

#######################
# Full example with app
#######################

library(iSEEu)
# Define ReducedDimensionPlot, VolcanoPlot, RowDataTable
# LinkedFeatureAssayPlot Panels for the main assay
rdp <- ReducedDimensionPlot()
vp <- VolcanoPlot()
rdt <- RowDataTable(RowSelectionSource = "VolcanoPlot1")
mlfap <- LinkedFeaturesAssayPlot(
SelectionExperiment = "(Main)",
LookupColumn = "(rowname)",
Experiment= "(Main)",
AltAssay = "proteins",
MapColumn = "(rowname)",
YAxisFeatureSource = "RowDataTable1",
PlotType = "Scatter + lines",
XAxis = "Column data",
XAxisColumnData = "sampleId"
)

# Define LinkedFeatureAssayPlot for peptides altExp assay  
alfap <- LinkedFeaturesAssayPlot(
YAxisFeatureSource = "RowDataTable1",
SelectionExperiment = "(Main)",
LookupColumn = "Proteins",
Experiment= "peptides",
AltAssay = "peptides_norm",
PlotType = "Scatter + lines",
XAxis = "Column data",
XAxisColumnData = "sampleId",
MapColumn = "Proteins"
)

# Run app 
data(sceProteinsPeptides)
app <- iSEE(
sceProteinsPeptides,
initial=list(rdp, vp, rdt, mlfap, alfap)
)

if (interactive()) {
shiny::runApp(app, launch.browser = TRUE)
}
### End Full example 
}

}
\seealso{
\code{\linkS4class{LinkedFeaturesAssayPlot}} for slot details.
}
