% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainExpPlot.R
\name{MainExpPlot}
\alias{MainExpPlot}
\title{Construct a MainExpPlot panel}
\usage{
MainExpPlot(...)
}
\arguments{
\item{...}{Named arguments corresponding to slots of
\code{\linkS4class{MainExpPlot}} or its parent classes.  Any slot not
supplied receives a sensible default: \code{PlotType} defaults to
\code{"Auto"}.}
}
\value{
A \code{\linkS4class{MainExpPlot}} object.
}
\description{
Creates an instance of the \code{\linkS4class{MainExpPlot}} class for use
as a panel in an iSEE application.
}
\examples{
\dontrun{
# Minimal panel using runtime defaults
MainExpPlot()

# Pre-configure to show scatter plot with lines connecting points
MainExpPlot(PlotType = "Scatter + lines")

# Pre-configure YAxisFeatureSource, PlotType, XAxis and XAxisColumnData
MainExpPlot(YAxisFeatureSource = "RowDataTable1", 
PlotType = "Scatter + lines", 
XAxis = "Column data",
XAxisColumnData = "sampleId")

# Example of iSEE instance with mainExp and altExp
# linked via the "Proteins" in rowData columns of MainExp and AltExp 
data("proteinsPeptidesSce")
iSEE(
sceProteinsPeptides,
initial = list(
RowDataTable(),
MainExpPlot(YAxisFeatureSource = "RowDataTable1", 
            PlotType = "Scatter + lines", 
            XAxis = "Column data",
            XAxisColumnData = "sampleId"),
AltExpPlot(YAxisFeatureSource = "RowDataTable1", 
           AltAssay = "peptides_norm",
           PlotType = "Scatter + lines", 
           XAxis = "Column data", 
           XAxisColumnData = "sampleId",
           LookupColumn = "Proteins",
           MapColumn = "Proteins")
)
)
}

}
\seealso{
\code{\linkS4class{MainExpPlot}} for a description of all slots.
}
